Move types from src/types/Scene.ts to ThreeDimensionalViz
These types are specific to the 3d panel and should live within the 3d panel folder.
